✓ Add all basic tools (Line triangle square rect and circle) https://app.asana.com/0/1202023384504214/1202023384504326/f Added the backend and ui for now we only have square,rect and triangle working but added the rest will only take adding new shaders and adding command for them.
✓Setup ViewPort for relative setup
have setup the viewPort and draw area now i can define the ortho viewport matrix to pan and zoom.
https://app.asana.com/0/1202023384504214/1202023384504321/f

Also updated the rendering part now there are separate classes that defines shader material mesh and model. now each shape is a model that is defined by the mesh shader and material.

Also added texture loading and tested texture loading to a shader with alpha. this will be useable with selection tool and setting up properties for the objects.

Added debug class and added debug window(this will be toggle able later for now it is fixed) will move all the logs to this.

+++ VectorPaint/VectorPaint/VectorPaint/TransformComponent.h
@@ -30,7 +30,7 @@ public:
 	}
 
 
-	TransformComponent(float p_x, float p_y, int p_w, int p_h, int p_scX, int p_scY)
+	TransformComponent(float p_x, float p_y, int p_w, int p_h, float p_scX, float p_scY)
 	{
 		Position.x = p_x;
 		Position.y = p_y;
